functionality for promp menu to add a payment type
