+++ app/controllers/customerCtrl.js
@@ -1,4 +1,4 @@
-'use strict';
+  'use strict';
 
 const prompt = require('prompt');
